refactor(tests): reorganize second-wave tests by module
Merge second-wave regression tests into the existing module-grouped
structure. Tests now live next to the module they cover rather than
being grouped by which batch created them.

Deleted 4 agent-grouped files, appended to 6 existing files, created
6 new module-specific files. test_defense_in_depth.py kept as
cross-cutting security tests.
